# modules/reservation/reservation-headless/reservation-headless-impl/rest-openapi.yaml
components:
  schemas:
...
    Purpose:
      description: Contains information about a room's designated purpose (meetings, townhall, etc.)
      properties:
        purposeId:
          description: The primary key for the Purpose entity.
          type: integer
          format: int64
          readOnly: true
        name:
          description: The purpose name.
          type: string
      required: 
        - name
      type: object
...
openapi: 3.0.1
info:
  description: "API for accessing Reservation details."
  license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  title: "Headless Reservation"
  version: v1.0
paths:
...
  "/purposes":
    get:
      operationId: getPurposesPage
      tags: ["Purpose"]
      description: Retrieve the list of purposes. Results can be paginated, filtered, searched and sorted.
      parameters:
        - in: query
          name: filter
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Purpose"
                type: array
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Purpose"
                type: array
    post:
      operationId: postPurpos
      tags: ["Purpose"]
      description: Create a new purpose tag.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Purpose"
          application/xml:
            schema:
              $ref: "#/components/schemas/Purpose"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purpose"
            application/xml:
              schema:
                $ref: "#/components/schemas/Purpose"
  "/purposes/{purposeId}":
    get:
      operationId: getPurposPurpose
      tags: ["Purpose"]
      description: Retrieves the purpose via its ID.
      parameters:
        - name: purposeId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purpose"
            application/xml:
              schema:
                $ref: "#/components/schemas/Purpose"
    put: 
      operationId: putPurposPurpose
      tags: ["Purpose"]
      description: Updates and overwrites the target purpose.
      parameters:
        - name: purposeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Purpose"
          application/xml:
            schema:
              $ref: "#/components/schemas/Purpose"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purpose"
            application/xml:
              schema:
                $ref: "#/components/schemas/Purpose"
    delete:
      operationId: deletePurposPurpose
      tags: ["Purpose"]
      description: Deletes the purpose and returns a 204 if the operation succeeds.
      parameters:
        - name: purposeId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: ""
          content:
            application/json: {}
            application/xml: {}
...